Got rid of json config serialization. Refactored BB library and plugin structure
