+++ internal/bitbucket/utils.go
